+++ src/constants/colorPalettes.ts
@@ -1,0 +1,68 @@
+export const COLOR_PALETTES = [
+  '#F9FAFB',
+  '#EBEEF2',
+  '#B2BECC',
+  '#788DA3',
+  '#53616F',
+  '#323941',
+  '#B32A2E',
+  '#A14217',
+  '#8E501A',
+  '#7E5815',
+  '#6D5F1B',
+  '#5C641D',
+  '#FD4649',
+  '#ED5C18',
+  '#CE7419',
+  '#B88011',
+  '#BCA51E',
+  '#859222',
+  '#FFA59B',
+  '#FFA87E',
+  '#FEAA5C',
+  '#F8AF20',
+  '#F2D41B',
+  '#B3C725',
+  '#FFE9E5',
+  '#FFEADE',
+  '#FFEAD6',
+  '#FFEBCD',
+  '#FFEF9B',
+  '#EFF2B0',
+  '#466929',
+  '#2C6C39',
+  '#206B51',
+  '#21696A',
+  '#22677D',
+  '#226492',
+  '#399E4F',
+  '#209D75',
+  '#219A9B',
+  '#2096B9',
+  '#1892D9',
+  '#028CFD',
+  '#84D046',
+  '#45D769',
+  '#10D69E',
+  '#0FD2D3',
+  '#02CDFE',
+  '#80C3FF',
+  '#DDF6C4',
+  '#CFF8D2',
+  '#C6F9E2',
+  '#CAF6F6',
+  '#D6F2FF',
+  '#E2EFFF',
+  '#4654C0',
+  '#7947AC',
+  '#A1347C',
+  '#787FF6',
+  '#AB6FE7',
+  '#EB47BD',
+  '#BAB5FB',
+  '#D3AEF3',
+  '#FB9FE1',
+  '#EEECFE',
+  '#F4EAFC',
+  '#FFE7F9',
+];
