+++ src/layout/RootLayout/RootLayout.tsx
@@ -1,11 +1,11 @@
 import { Outlet, useLocation, useNavigate } from 'react-router-dom';
-import * as S from './style';
-import { GUEST_UUID, LOGIN, MY_REGION } from '@/constants/localStorage/key';
+import * as S from './RootLayout.style';
+import { GUEST_UUID, LOGIN, MY_REGION } from '@/constants/localStorageKey';
 import { useMutation } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 import { guestLogin } from '@/services/login';
 import useAppDispatch from '@/hooks/useAppDispatch';
-import A2hsSnackbar from './components/A2hsSnackbar';
+import A2hsSnackbar from './A2hsSnackbar/A2hsSnackbar';
 import { goelocationActions } from '@/store/slice/geolocationSlice';
 import useGeolocation from '@/hooks/useGeolocation';
 import { storeAccessToken, storeUser } from '@/utils/auth';
