+++ src/pages/User/TabSection/LookbookList/LookbookCard/style.ts
@@ -1,4 +1,4 @@
-import { LOOKBOOK_WEATHER_TYPE } from '@/constants/Lookbook/data';
+import { LOOKBOOK_WEATHER_TYPE } from '@/constants/lookbook';
 import { WeatherType } from '@/types/weather';
 import styled from '@emotion/styled';
 import { IconButton } from '@mui/material';
