+++ src/pages/UserLookbookCreate/Edit/ColorPalette/ColorPalette.style.ts
@@ -1,4 +1,4 @@
-import { MAX_WIDTH } from '@/constants/css/value';
+import { MAX_WIDTH } from '@/constants/css';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
