+++ src/pages/UserLookbookCreate/WeatherHeadline/WeatherHeadline.tsx
@@ -1,10 +1,10 @@
-import { LOOKBOOK_WEATHER_TYPE } from '@/constants/Lookbook/data';
 import { WeatherType } from '@/types/weather';
 import { S } from './WeatherHeadline.style';
 import { useLocation } from 'react-router-dom';
 import { LocationState } from '../UserLookbookCreatePage';
 import { TempCondition } from '@/pages/Home/ClothesSection/ClothesSection';
 import { memo } from 'react';
+import { LOOKBOOK_WEATHER_TYPE } from '@/constants/lookbook';
 
 type WeatherHeadlineProps = {
   weatherType: WeatherType;
